support component && optimize manifest
